Replaced interact with jump
